Split the test_utils module into a package.
The module was getting pretty long and most of the things were only
loosely related to each other.
