+++ utils/test_utils/xvfb.py
@@ -1,0 +1,33 @@
+# Amara, universalsubtitles.org
+#
+# Copyright (C) 2015 Participatory Culture Foundation
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as
+# published by the Free Software Foundation, either version 3 of the
+# License, or (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU Affero General Public License for more details.
+#
+# You should have received a copy of the GNU Affero General Public License
+# along with this program. If not, see
+# http://www.gnu.org/licenses/agpl-3.0.html.
+
+from xvfbwrapper import Xvfb
+
+_xvfb = None
+def start_xvfb():
+    global _xvfb
+    if _xvfb is None:
+        _xvfb = Xvfb(width=1920, height=1080)
+        _xvfb.start()
+
+def stop_xvfb():
+    global _xvfb
+    if _xvfb is not None:
+        _xvfb.stop()
+        _xvfb = None
+
